src/requirements.txt
azure-functions>=1.18.0
yfinance>=0.2.18
requests>=2.31.0
